+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [v0.3.0](https://github.com/naacl2022-reproducibility-track/naacl-utils/releases/tag/v0.3.0) - 2022-02-22
+
+### Changed
+
+- `naacl-utils verify` now takes a path for the expected output, instead of a string.
+
+## [v0.2.0](https://github.com/naacl2022-reproducibility-track/naacl-utils/releases/tag/v0.2.0) - 2022-01-19
+
+### Added
+
+- Added `naacl-utils verify` command.
+
 ## [v0.1.5](https://github.com/naacl2022-reproducibility-track/naacl-utils/releases/tag/v0.1.5) - 2021-12-14
 
 ### Added